Move plugin utils to dedicated module
